+++ app/src/main/java/ayds/newyork/songinfo/moredetails/model/data/external/nytimes/artistinfo/NYTimesToArtistInfoResolver.kt
@@ -1,0 +1,2 @@
+package ayds.newyork.songinfo.moredetails.model.data.external.nytimes.artistinfo
+
