+++ app/src/main/java/ayds/newyork/songinfo/moredetails/model/data/local/nytimes/NYTimesArtistInfoLocalStorage.kt
@@ -1,0 +1,4 @@
+package ayds.newyork.songinfo.moredetails.model.data.local.nytimes
+
+interface NYTimesArtistInfoLocalStorage {
+}
